Add option to define phase by duration and/or distance.
Define EndType, add property that defines Phase end, also property for what time or distance remains in a phase.  Add a running field to track when the phase is running.  Use timeRemaining field to note when phase or workout is paused.

Move ding at an end of the phase to MainViewController, create protocol for MainViewController to implement in order to process phase changes and the end of the workout.